Use CurrentStatus in Timeline
