+++ src/js/claims-status/containers/AppealsV2StatusPage.jsx
@@ -3,9 +3,9 @@ import PropTypes from 'prop-types';
 import { connect } from 'react-redux';
 
 import { getAppealsV2 } from '../actions/index.jsx';
+import { getStatusContents } from '../utils/appeals-v2-helpers';
 
 import Timeline from '../components/appeals-v2/Timeline';
-import CurrentStatus from '../components/appeals-v2/CurrentStatus';
 import Alerts from '../components/appeals-v2/Alerts';
 import WhatsNext from '../components/appeals-v2/WhatsNext';
 import Docket from '../components/appeals-v2/Docket';
@@ -18,17 +18,18 @@ class AppealsV2StatusPage extends React.Component {
   componentDidMount() {
     // Make sure we grab the appeals if we don't have them already
     // Useful if the user goes directly to the appeal status without going to the list first
-    if (!this.props.appeal) {
+    if (this.props.appeal === AppealsV2StatusPage.defaultProps.appeal) {
       this.props.getAppealsV2();
     }
   }
 
   render() {
-    const events = this.props.appeal ? this.props.appeal.attributes.events : [];
+    const { events, status } = this.props.appeal.attributes;
+    const { type, details } = status;
+    const currentStatus = getStatusContents(type, details);
     return (
       <div>
-        <Timeline events={events}/>
-        <CurrentStatus/>
+        <Timeline events={events} currentStatus={currentStatus}/>
         <Alerts/>
         <WhatsNext/>
         <Docket/>
@@ -36,6 +37,20 @@ class AppealsV2StatusPage extends React.Component {
     );
   }
 }
+
+AppealsV2StatusPage.defaultProps = {
+  appeal: {
+    id: '',
+    type: '',
+    attributes: {
+      events: [],
+      status: {
+        type: '',
+        details: {}
+      }
+    }
+  }
+};
 
 AppealsV2StatusPage.propTypes = {
   params: PropTypes.shape({
